Refactor encoding mod
