+++ aws-lc-rs/tests/ecdsa_tests.rs
@@ -3,8 +3,7 @@
 // Modifications copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR ISC
 
-use aws_lc_rs::encoding::AsBigEndian;
-use aws_lc_rs::signature::EcPrivateKeyRfc5915Der;
+use aws_lc_rs::encoding::{AsBigEndian, EcPrivateKeyRfc5915Der};
 use aws_lc_rs::{
     encoding::AsDer,
     rand::SystemRandom,
